add: lay calories ben progress, lay latest calories, lay today workout

+++ src/main/java/com/example/gymapp/dto/request/HistoryRequest.java
@@ -24,7 +24,7 @@ public class HistoryRequest {
     @NotNull(message = "progression.value.null")
     @Min(0)
     @Max(1000)
-    private double calories;
+    private float calories;
 
     private LocalDate createdDate;
